refactor(core): move context services to packages/core/src/context/
