+++ packages/core/src/context/embeddingService.test.ts
