+++ packages/core/src/context/embeddingService.ts
